pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bubble-memory"
version = "0.2.0"
description = "Event Sourcing based belief formation system for long-term AI agent memory"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AutismAccelerator", email = "your.email@example.com"}
]
keywords = ["ai", "memory", "agent", "graph", "llm", "vector-store"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
# Core runtime deps — provider SDKs are optional extras below.
dependencies = [
    "falkordb>=1.6.0",
    "httpx>=0.28.0",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]

[project.optional-dependencies]
# LLM providers — install the one(s) you need.
anthropic = ["anthropic>=0.86.0"]
openai    = ["openai>=1.0"]
gemini    = ["google-generativeai>=0.8"]

# Install all providers at once:
all-providers = [
    "anthropic>=0.86.0",
    "openai>=1.0",
    "google-generativeai>=0.8",
]

# Development tools:
dev = ["build", "twine", "ruff", "anthropic>=0.86.0"]

[project.urls]
Homepage   = "https://github.com/AutismAccelerator/bubble"
Repository = "https://github.com/AutismAccelerator/bubble"
Issues     = "https://github.com/AutismAccelerator/bubble/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 230